+++ Turret_2.vcxproj
@@ -73,21 +73,13 @@
     <ClInclude Include="src\game\generation\perlin_noise.hpp" />
     <ClInclude Include="src\game\generation\hash_noise.hpp" />
     <ClInclude Include="src\game\generation\spot_generator.hpp" />
-    <ClInclude Include="src\game\physics\mob.hpp" />
     <ClInclude Include="src\game\physics\mobs_system.hpp" />
-    <ClInclude Include="src\game\physics\mob_ai.hpp" />
     <ClInclude Include="src\game\physics\mob_soa.hpp" />
-    <ClInclude Include="src\game\physics\shells_system.hpp" />
     <ClInclude Include="src\game\physics\physics_base.hpp" />
-    <ClInclude Include="src\game\physics\shell.hpp" />
     <ClInclude Include="src\game\physics\team\team.hpp" />
     <ClInclude Include="src\game\physics\team\teams_pool.hpp" />
-    <ClInclude Include="src\game\physics\turret.hpp" />
-    <ClInclude Include="src\game\physics\turret_types.hpp" />
     <ClInclude Include="src\game\player\camera.hpp" />
     <ClInclude Include="src\game\player\player_controller.hpp" />
-    <ClInclude Include="src\game\presets\mob_presets.hpp" />
-    <ClInclude Include="src\game\presets\shell_presets.hpp" />
     <ClInclude Include="src\game\script_libs\script_libs.hpp" />
     <ClInclude Include="src\game\weather\weather.hpp" />
     <ClInclude Include="src\game\world\world.hpp" />
@@ -155,7 +147,6 @@
     <ClCompile Include="src\game\frontend\fr_world_saving.cpp" />
     <ClCompile Include="src\game\generation\generation.cpp" />
     <ClCompile Include="src\game\physics\mobs_system.cpp" />
-    <ClCompile Include="src\game\physics\shells_system.cpp" />
     <ClCompile Include="src\game\physics\team\team.cpp" />
     <ClCompile Include="src\game\player\camera.cpp" />
     <ClCompile Include="src\game\player\player_controller.cpp" />
