+++ src/game/physics/mob_soa.hpp
@@ -80,67 +80,3 @@ inline void removeMob(MobSoA& soa, size_t index) {
     soa.preset.pop_back();
     soa.hitbox.pop_back();
 }
-
-static t1_finline void move(MobSoA& soa, const size_t index, const PixelCoord vector) {
-    soa.position[index] = soa.position[index] + vector;
-    soa.hitbox[index].move(vector);
-}
-
-static inline void resolveCollisions(MobSoA& soa, const size_t current, const size_t end) {
-    for (size_t other = current + 1; other < end; ++other) {
-        if (!soa.hitbox[current].intersects(soa.hitbox[other])) continue;
-
-        const PixelCoord overlap = soa.hitbox[current].overlap(soa.hitbox[other]);
-
-        const bool pushX = (overlap.x < overlap.y);
-        const float distance = pushX ? overlap.x : overlap.y;
-
-        const float pushDirection = pushX ? (soa.position[current].x > soa.position[other].x ? 1.0f : -1.0f)
-            : (soa.position[current].y > soa.position[other].y ? 1.0f : -1.0f);
-
-        const float totalRadius = soa.preset[current]->hitboxRadius + soa.preset[other]->hitboxRadius;
-        const float invertedTotalRadius = 1.0f / totalRadius;
-        const float mob1Weight = soa.preset[other]  ->hitboxRadius * invertedTotalRadius;
-        const float mob2Weight = soa.preset[current]->hitboxRadius * invertedTotalRadius;
-
-        PixelCoord pushVec;
-        if (pushX) pushVec.x = distance * pushDirection;
-        else /*Y*/ pushVec.y = distance * pushDirection;
-
-        move(soa, current, pushVec * mob1Weight);
-        move(soa, other,   pushVec * -mob2Weight);
-    }
-}
-
-static inline void moveByAI(MobSoA& soa, const size_t index) {
-    if (false)
-        return;
-    //updateAI
-    PixelCoord AIVector; //getFromAI
-    Angle AIAngle = 0; //getFromAI
-    soa.velocity[index] = AIVector * soa.preset[index]->maxSpeed;
-    soa.angle[index] = AIAngle;
-    move(soa, index, soa.velocity[index]);
-}
-
-inline void processMobs(MobSoA& soa) {
-    size_t end = soa.id.size();
-    for (size_t i = 0; i < end; ++i) {
-        resolveCollisions(soa, i, end);
-    }
-}
-
-inline void drawMobs(const MobSoA& soa, const Camera& camera, const float tickOfset) {
-    Sprite sprite;
-    for (size_t i = 0; i < soa.id.size(); ++i) {
-        if (!camera.contains(t1::tile(soa.position[i])))
-            return;
-        auto& visual = soa.preset[i]->visual;
-        sprite.setTexture(*visual.texture);
-        sprite.setOrigin(visual.origin);
-        sprite.setSize(visual.size);
-        sprite.setPosition(soa.position[i] + soa.velocity[i] * tickOfset);
-        sprite.setRotationRad(soa.angle[i]);
-        sprite.draw();
-    }
-}
