+++ src/game/physics/mobs_system.hpp
@@ -1,11 +1,9 @@
 #pragma once
-#include <list>
 
-struct Mob;
+struct MobSoA;
 class Camera;
-class TeamsPool;
 
 namespace mobs {
-    void processMobs(std::list<Mob>& mobs, TeamsPool& teams);
-    void drawMobs(std::list<Mob>& mobs, const Camera& camera, const float tickOfset);
+    void processMobs(MobSoA& soa);
+    void drawMobs(const MobSoA& soa, const Camera& camera, const float tickOfset);
 }
