--- src/game/physics/shells_system.hpp
@@ -1,11 +1,0 @@
-#pragma once
-#include <list>
-
-struct Shell;
-class Camera;
-class TeamsPool;
-
-namespace shells {
-    void processShells(std::list<Shell>& shells, TeamsPool& teams);
-    void drawShells(std::list<Shell>& shells, const Camera& camera, const float tickOfset);
-}
