+++ src/game/world_drawer/entities_drawer.cpp
@@ -2,8 +2,8 @@
 //
 #include "game/physics/team/teams_pool.hpp"
 #include "game/world/world.hpp"
-#include "game/physics/mob_soa.hpp"
+#include "game/physics/mobs_system.hpp"
 
 void EntitiesDrawer::draw(const Camera& camera, const MobSoA& soa, const float tickOfset) {
-    drawMobs(soa, camera, tickOfset);
+    mobs::drawMobs(soa, camera, tickOfset);
 }
